fix:跑通api-createNewTask
主键使用 BIGINT（Long）

Task和它的info之间需要互相设置-->TaskCreateReq的toTaskDTO方法

+++ src/main/java/com/zmark/mytodo/controller/TaskController.java
@@ -26,7 +26,7 @@ public class TaskController {
     }
 
     @GetMapping("/api/task/get-simple-info/{task-id}")
-    public Result getSimpleInfoById(@PathVariable("task-id") int taskId) {
+    public Result getSimpleInfoById(@PathVariable("task-id") Long taskId) {
         try {
             TaskSimpleResp respData = TaskSimpleResp.from(taskService.findTaskById(taskId));
             return ResultFactory.buildSuccessResult(respData);
@@ -37,7 +37,7 @@ public class TaskController {
     }
 
     @GetMapping("/api/task/get-detail-info/{task-id}")
-    public Result getDetailInfoById(@PathVariable("task-id") int taskId) {
+    public Result getDetailInfoById(@PathVariable("task-id") Long taskId) {
         try {
             TaskDetailResp respData = TaskDetailResp.from(taskService.findTaskById(taskId));
             return ResultFactory.buildSuccessResult(respData);
