+++ src/main/java/com/zmark/mytodo/dao/MyDayTaskDAO.java
@@ -3,5 +3,5 @@ package com.zmark.mytodo.dao;
 import com.zmark.mytodo.entity.MyDayTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MyDayTaskDAO extends JpaRepository<MyDayTask, Integer> {
+public interface MyDayTaskDAO extends JpaRepository<MyDayTask, Long> {
 }
