+++ src/main/java/com/zmark/mytodo/dao/StepInfoDAO.java
@@ -3,5 +3,5 @@ package com.zmark.mytodo.dao;
 import com.zmark.mytodo.entity.StepInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StepInfoDAO extends JpaRepository<StepInfo, Integer> {
+public interface StepInfoDAO extends JpaRepository<StepInfo, Long> {
 }
