+++ src/main/java/com/zmark/mytodo/dao/TaskContentInfoDAO.java
@@ -3,5 +3,5 @@ package com.zmark.mytodo.dao;
 import com.zmark.mytodo.entity.TaskContentInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskContentInfoDAO extends JpaRepository<TaskContentInfo, Integer> {
+public interface TaskContentInfoDAO extends JpaRepository<TaskContentInfo, Long> {
 }
