+++ src/main/java/com/zmark/mytodo/dao/TaskDAO.java
@@ -3,6 +3,6 @@ package com.zmark.mytodo.dao;
 import com.zmark.mytodo.entity.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskDAO extends JpaRepository<Task, Integer> {
-    Task findTaskById(int id);
+public interface TaskDAO extends JpaRepository<Task, Long> {
+    Task findTaskById(Long id);
 }
