+++ src/main/java/com/zmark/mytodo/dao/TaskPriorityInfoDAO.java
@@ -3,5 +3,5 @@ package com.zmark.mytodo.dao;
 import com.zmark.mytodo.entity.TaskPriorityInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskPriorityInfoDAO extends JpaRepository<TaskPriorityInfo, Integer> {
+public interface TaskPriorityInfoDAO extends JpaRepository<TaskPriorityInfo, Long> {
 }
