+++ src/main/java/com/zmark/mytodo/dao/TaskTimeInfoDAO.java
@@ -3,5 +3,5 @@ package com.zmark.mytodo.dao;
 import com.zmark.mytodo.entity.TaskTimeInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskTimeInfoDAO extends JpaRepository<TaskTimeInfo, Integer> {
+public interface TaskTimeInfoDAO extends JpaRepository<TaskTimeInfo, Long> {
 }
